+++ demo/src/CodeSandbox.js
@@ -3,7 +3,7 @@ import React from "react";
 export default ({ codeSandbox, style }) => {
   if (!codeSandbox) {
     return "";
-  }console.log(codeSandbox);
+  }
   return (
     <a style={style} href={"https://codesandbox.io/s/"+codeSandbox+"?fontsize=14&hidenavigation=1&theme=dark"}>
       <img