+++ demo/src/Range.js
@@ -2,7 +2,7 @@ import React from "react";
 import Example from "./Example.js";
 
 const code = `<HighlightWithinTextarea
-  highlight={[2, 6]}
+  highlight={[6, 23]}
   ...
 />`;
 
@@ -11,10 +11,10 @@ const Range = () => {
     <>
       <Example
         title="Array of Two Numbers (Range)"
-        text="An array of exactly two numbers is treated as a range. Highlighting starts at the first character index (inclusive) and ends at the second character index (exclusive)."
+        text="An array of exactly two numbers is treated as a range. Highlighting starts at the first character index (inclusive) and ends at the second character index (exclusive).  Ranges can also span multiple lines with the line break character (or characters for Windows) being included"
         initialValue="apple watermelon banana orange mango"
         code={code}
-        highlight={[2, 5]}
+        highlight={[6, 23]}
         codeSandbox="rhwta-range-l85l2"
       />
     </>